implemented Endpoint scope constraints

+++ laserminipod_server/lib/src/endpoints/handle_endpoint.dart
@@ -1,4 +1,5 @@
 import 'package:laserminipod_server/src/di.dart';
+import 'package:laserminipod_server/src/extensions/endpoint_ext.dart';
 import 'package:laserminipod_server/src/domain/handle_service.dart';
 import 'package:serverpod/serverpod.dart';
 import '../generated/protocol.dart';
@@ -12,15 +13,18 @@ class HandleEndpoint extends Endpoint {
 
   /// returns id of created handle
   Future<int> addHandle(Session session, int x, int y, int radius) async {
+    await validateAdminScope(session);
     return await _handleService.addHandle(session, x, y, radius);
   }
 
   Future<void> editHandle(
       Session session, int id, int x, int y, int radius) async {
+    await validateAdminScope(session);
     await _handleService.editHandle(session, id, x, y, radius);
   }
 
   Future<void> removeHandle(Session session, int id) async {
+    await validateAdminScope(session);
     await _handleService.removeHandle(session, id);
   }
 }
